fix: move sse.py out of vibetuner.frontend to avoid circular imports
Importing from vibetuner.frontend.sse triggers vibetuner.frontend.__init__
which calls load_app_config() at module level, causing circular imports
when user route files import SSE utilities.

Move the SSE module to vibetuner/sse.py (top-level package) and replace
the original with a backward-compatible re-export shim.

Closes #1001
